chore: remove orchestrators
